requirements.txt
anthropic
feedparser
requests
python-dotenv
pyyaml
